configure: fail configure step without cuda
Make the CUDA check m4 snippet cry and fail if CUDA or related
dependencies are not found. Without enable_cuda explicitly set to yes,
the macro only emits a warning, letting configure finish successfully.
As this project requires CUDA in any case, that results in a build
error, to which a simple configure failure is preferred.

Change-Id: Ia1ca0e50b05c204f71e1ee8fb085bc0b9825f44a
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6290
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>