configure.ac: Use g++ for plain c++, and move to AX_PROG_NVCC
Replace AX_CHECK_CUDA with the custom, minimal macro AX_PROG_NVCC.

The new macro defines NVCC and NVCC_CXXFLAGS as precious variables that
can be overridden, and does proper target arch detection rather than
guessing from hard-coded paths in the CUDA compiler directory.

Setting CFLAGS and LIBS for cuda is left to to pkg-config.

Also, this changes the compiler for standard, non-CUDA c++ source files
from nvcc to g++, or whatever CXX resolves to. This overcomes the
limited set of compiler flags that is supported by nvcc.

Since linking the final executable with the gpu "device code" does
require nvcc, a Makefile rule is added to ensure nvcc is used for the
final link.

Change-Id: Ib5545c71fa21dbc874f0fdce2f1f40664eb2784a
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6396
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>