Properly count the number of buffers initialized and fix the error path
The loop that initialize the buffers was setting the buffer index as
the number of buffer, offsetting it by one. As such the last buffer
wasn't used for the capture and wasn't properly released on close.

Furthermore the buffer init error path wasn't unmapping the buffers
that had already been mapped. After adding the missing munmap in the
error path we also need to store the original error before jumping in
the error path as munmap might also set errno.

Change-Id: I893c89c767307218a326ec635a9372d3bd0e465d
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7570
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>